feat: add controller class
